# clib.Session: Refactor the __getitem__ special method to avoid calling API function GMT_Get_Enum repeatedly